Added middleware for the admin routes

+++ src/app/api/admin/delete/route.js
@@ -1,5 +1,6 @@
 import connectDB from "@/DB"
 import Product from "@/Models/product"
+import AuthUser from "@/middleware/AuthUser"
 import { NextResponse } from "next/server"
 
 
@@ -9,13 +10,15 @@ export const DELETE = async (req) => {
     try {
         await connectDB()
 
-        const {searchParams} = new URL(req.url)
-        const id = searchParams.get('id')
-        if(!id) 
-            return NextResponse.json({
-                success: false,
-                message: 'Product ID is required'
-            })
+        const isAuth = await AuthUser(req)
+        if(isAuth?.role === 'admin') {
+            const {searchParams} = new URL(req.url)
+            const id = searchParams.get('id')
+            if(!id) 
+                return NextResponse.json({
+                    success: false,
+                    message: 'Product ID is required'
+                })
         
             const deletedProduct = await Product.findByIdAndDelete(id)
             if (deletedProduct)
@@ -28,6 +31,13 @@ export const DELETE = async (req) => {
                     success: false,
                     message: 'Failed to delete product.'
                 })
+        } else {
+            return NextResponse.json({
+                success: false,
+                message: 'You are not authorized.'
+            })
+        }
+
     } catch (err) {
         console.log(err)
         return NextResponse.json({
